rename libquix-lexer to libnitrate-lexer
