+++ libnitrate-lexer/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.15)
-project(libquix-lexer CXX)
+project(libnitrate-lexer CXX)
 
 #================== CONFIGURE BUILD ==================
 set(CMAKE_CXX_STANDARD 20)
@@ -13,7 +13,7 @@ string(REGEX REPLACE " .*" "" SHA256SUMS ${SHA256SUMS})
 string(TOUPPER ${SHA256SUMS} SHA256SUMS)
 string(SUBSTRING ${SHA256SUMS} 0 32 SHA256SUMS)
 set(__TARGET_VERSION "\"QLEX_${SHA256SUMS}\"")
-message(STATUS "libquix-lexer VersionId: ${__TARGET_VERSION}")
+message(STATUS "libnitrate-lexer VersionId: ${__TARGET_VERSION}")
 
 #================== SET BUILD FLAGS ==================
 if(CMAKE_BUILD_TYPE STREQUAL "Debug")
@@ -27,10 +27,10 @@ endif()
 
 file(GLOB_RECURSE CXX_SOURCES "src/*.cc")
 
-add_library(quix-lexer ${CXX_SOURCES})
-target_include_directories(quix-lexer PUBLIC src "include" ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-core/include)
-target_link_libraries(quix-lexer PUBLIC nitrate-core)
-add_dependencies(quix-lexer nitrate-core)
+add_library(nitrate-lexer ${CXX_SOURCES})
+target_include_directories(nitrate-lexer PUBLIC src "include" ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-core/include)
+target_link_libraries(nitrate-lexer PUBLIC nitrate-core)
+add_dependencies(nitrate-lexer nitrate-core)
 
-install(TARGETS quix-lexer DESTINATION lib)
+install(TARGETS nitrate-lexer DESTINATION lib)
 install(DIRECTORY include/ DESTINATION "include")
