+++ libnitrate-lexer/include/nitrate-lexer/Token.h
