+++ libquix-parser/CMakeLists.txt
@@ -30,9 +30,9 @@ file(GLOB_RECURSE CXX_SOURCES "src/*.cc")
 add_library(quix-parser ${CXX_SOURCES})
 target_include_directories(quix-parser PUBLIC src "include"
   ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-core/include
-  ${CMAKE_CURRENT_SOURCE_DIR}/../libquix-lexer/include)
-target_link_libraries(quix-parser PUBLIC nitrate-core quix-lexer deflate)
-add_dependencies(quix-parser nitrate-core quix-lexer)
+  ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-lexer/include)
+target_link_libraries(quix-parser PUBLIC nitrate-core nitrate-lexer deflate)
+add_dependencies(quix-parser nitrate-core nitrate-lexer)
 
 install(TARGETS quix-parser DESTINATION lib)
 install(DIRECTORY include/ DESTINATION "include")
