+++ libquix-parser/include/quix-parser/Config.h
@@ -36,7 +36,7 @@
 #include <initializer_list>
 #endif
 
-#include <quix-lexer/Lexer.h>
+#include <nitrate-lexer/Lexer.h>
 #include <stdbool.h>
 
 #ifdef __cplusplus
