+++ libquix-parser/src/Serialize.cc
@@ -33,7 +33,7 @@
 
 #include <libdeflate.h>
 #include <nitrate-core/Error.h>
-#include <quix-lexer/Lexer.h>
+#include <nitrate-lexer/Lexer.h>
 #include <quix-parser/Node.h>
 
 #include <ParserStruct.hh>
