+++ libquix-parser/src/parser/FunctionParser.cc
@@ -32,8 +32,8 @@
 /// TODO: Source location
 #include <parser/Parse.h>
 
-#include "quix-lexer/Lexer.h"
-#include "quix-lexer/Token.h"
+#include "nitrate-lexer/Lexer.h"
+#include "nitrate-lexer/Token.h"
 
 using namespace qparse;
 using namespace qparse::parser;
