+++ libquix/src/Transform.cc
@@ -32,9 +32,9 @@
 #define LIBQUIX_INTERNAL
 #include <nitrate-core/Env.h>
 #include <nitrate-core/Lib.h>
+#include <nitrate-lexer/Lib.h>
 #include <quix-codegen/Lib.h>
 #include <quix-ir/Lib.h>
-#include <quix-lexer/Lib.h>
 #include <quix-parser/Lib.h>
 #include <quix-prep/Lib.h>
 #include <quix/code.h>
